+++ contract-tests/sse-contract-tests/src/entity_manager.cpp
@@ -27,7 +27,7 @@ std::optional<std::string> EntityManager::create(ConfigParams const& params) {
     }
 
     if (params.body) {
-        client_builder.body(std::move(*params.body));
+        client_builder.body(*params.body);
     }
 
     if (params.readTimeoutMs) {
@@ -35,16 +35,22 @@ std::optional<std::string> EntityManager::create(ConfigParams const& params) {
             std::chrono::milliseconds(*params.readTimeoutMs));
     }
 
+    if (params.initialDelayMs) {
+        client_builder.initial_reconnect_delay(
+            std::chrono::milliseconds(*params.initialDelayMs));
+    }
+
     client_builder.logger([this](std::string msg) {
         LD_LOG(logger_, LogLevel::kDebug) << std::move(msg);
     });
 
-    client_builder.receiver([copy = poster](launchdarkly::sse::Event e) {
-        copy->post_event(std::move(e));
+    client_builder.receiver([copy = poster](launchdarkly::sse::Event event) {
+        copy->post_event(std::move(event));
     });
 
-    client_builder.errors(
-        [copy = poster](launchdarkly::sse::Error e) { copy->post_error(e); });
+    client_builder.errors([copy = poster](launchdarkly::sse::Error event) {
+        copy->post_error(event);
+    });
 
     auto client = client_builder.build();
     if (!client) {
@@ -53,7 +59,7 @@ std::optional<std::string> EntityManager::create(ConfigParams const& params) {
         return std::nullopt;
     }
 
-    client->run();
+    client->async_connect();
 
     entities_.emplace(id, std::make_pair(client, poster));
     return id;